+++ src/cg/gui/LotsHeldTable.java
@@ -37,7 +37,7 @@ public class LotsHeldTable extends JTable {
 		 * SharesHeldTableRenderer r = new SharesHeldTableRenderer();
 		 * setDefaultRenderer(Object.class, r); setDefaultRenderer(Float.class,
 		 * r); setDefaultRenderer(Integer.class, r);
-		 * setDefaultRenderer(Long.class, r);
+		 * setDefaultRenderer(Integer.class, r);
 		 */
 	}
 
@@ -80,7 +80,7 @@ public class LotsHeldTable extends JTable {
 	}
 
 	private void updateSharesHeld() {
-		Long shares = model.getSharesHeld(rowFilter);
+		Integer shares = model.getSharesHeld(rowFilter);
 		sharesHeld.setValue(shares);
 	}
