+++ src/cg/gui/TradeOpenPositionsFilter.java
@@ -6,7 +6,7 @@ public class TradeOpenPositionsFilter extends RowFilter<TradeTableModel,Integer>
 
 	   public boolean include(Entry<? extends TradeTableModel, ? extends Integer> entry) {
 		   if( entry.getStringValue(2).startsWith("Buy") ){ // if its a buy trade
-			   Long sharesHeld = (Long)entry.getValue(5);
+			   Integer sharesHeld = (Integer)entry.getValue(5);
 			   if( sharesHeld > 0 )  // and some shares are held
 				   return true;
 			   else
